fix: Replace Probot Auto-merge with Merge Pal
Functionality is the same with the difference that a PR against
non-standard base branched would also get performed when the `merge`
label is set.

The previous config contained blocked branches:

```
blockingBaseBranches:
  - regex: ^feature/
  - regex: ^hotfix/
  - regex: ^bug/
```

Since PRs against any of these kind of branches should not happen with
trunk-based development and there is the security measure that you have
to add the `merge` label, I did not try to reproduce this behaviour with
Merge Pal.